requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-extensions==3.2.3
djangorestframework==3.14.0
filelock==3.13.1
fsspec==2023.10.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
panda==0.3.1
pandas==2.1.2
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
